+++ geotrek/maintenance/tests/test_models.py
@@ -6,6 +6,7 @@ from django.contrib.contenttypes.models import ContentType
 from django.contrib.gis.geos import LineString, Point
 from django.test import TestCase
 
+from geotrek.core.models import Topology
 from geotrek.core.tests.factories import (
     PathFactory,
     StakeFactory,
@@ -28,6 +29,7 @@ from geotrek.maintenance.tests.factories import (
     SignageInterventionFactory,
 )
 from geotrek.outdoor.tests.factories import CourseFactory, SiteFactory
+from geotrek.signage.models import Signage
 from geotrek.signage.tests.factories import BladeFactory, SignageFactory
 
 
@@ -339,6 +341,45 @@ class InterventionTest(TestCase):
     def test_target_display_csv_none(self):
         interv = InterventionFactory.create(target=None)
         self.assertIn("-", interv.target_csv_display)
+
+    def test_duplication_with_topology_target(self):
+        """
+        Test that duplicating an intervention with a topology target creates an independent topology copy.
+        https://github.com/GeotrekCE/Geotrek-admin/issues/3845
+        """
+        topology = TopologyFactory.create()
+        intervention = InterventionFactory.create(
+            target=topology,
+            target_id=topology.pk,
+            target_type=ContentType.objects.get_for_model(Topology),
+            name="intervention test",
+        )
+        self.assertEqual(Intervention.objects.count(), 1)
+        intervention_copy = Intervention.duplicate(intervention)
+        self.assertEqual(Intervention.objects.count(), 2)
+        self.assertEqual(intervention_copy.target_type, intervention.target_type)
+        self.assertNotEqual(intervention_copy.target_id, intervention.target_id)
+        topology_copy = intervention_copy.target
+        self.assertEqual(topology_copy.geom, topology.geom)
+        self.assertEqual(topology_copy.offset, topology.offset)
+        self.assertEqual(topology_copy.kind, "INTERVENTION")
+
+    def test_duplication_without_topology_target(self):
+        """
+        Test that duplicating an intervention without topology target duplicate the target id.
+        """
+        signage = SignageFactory.create()
+        intervention = InterventionFactory.create(
+            target=signage,
+            target_id=signage.pk,
+            target_type=ContentType.objects.get_for_model(Signage),
+            name="intervention test",
+        )
+        self.assertEqual(Intervention.objects.count(), 1)
+        intervention_copy = Intervention.duplicate(intervention)
+        self.assertEqual(Intervention.objects.count(), 2)
+        self.assertEqual(intervention_copy.target_type, intervention.target_type)
+        self.assertEqual(intervention_copy.target_id, intervention.target_id)
 
 
 @skipIf(not settings.TREKKING_TOPOLOGY_ENABLED, "Test with dynamic segmentation only")