+++ docs/changelog.rst
@@ -5,6 +5,10 @@ CHANGELOG
 2.126.1+dev     (XXXX-XX-XX)
 ----------------------------
 
+**Bug fixes**
+
+* Fix intervention duplication to also duplicate its target when linked to its own topology instead of another object (e.g. signage, etc) (refs #3845)
+
 
 2.126.1         (2026-08-14)
 ----------------------------